add flag to remove disk db before running

+++ cmd/harmony.go
@@ -48,11 +48,13 @@ func attackDetermination(attackedMode int) bool {
 }
 
 // InitLDBDatabase initializes a LDBDatabase.
-func InitLDBDatabase(ip string, port string) (*ethdb.LDBDatabase, error) {
+func InitLDBDatabase(ip string, port string, freshDB bool) (*ethdb.LDBDatabase, error) {
 	dbFileName := fmt.Sprintf("./db/harmony_%s_%s", ip, port)
-	var err = os.RemoveAll(dbFileName)
-	if err != nil {
-		fmt.Println(err.Error())
+	if freshDB {
+		var err = os.RemoveAll(dbFileName)
+		if err != nil {
+			fmt.Println(err.Error())
+		}
 	}
 	return ethdb.NewLDBDatabase(dbFileName, 0, 0)
 }
@@ -82,6 +84,7 @@ func main() {
 	logFolder := flag.String("log_folder", "latest", "the folder collecting the logs of this execution")
 	attackedMode := flag.Int("attacked_mode", 0, "0 means not attacked, 1 means attacked, 2 means being open to be selected as attacked")
 	dbSupported := flag.Bool("db_supported", true, "false means not db_supported, true means db_supported")
+	freshDB := flag.Bool("fresh_db", false, "true means the existing disk based db will be removed")
 	profile := flag.Bool("profile", false, "Turn on profiling (CPU, Memory).")
 	metricsReportURL := flag.String("metrics_report_url", "", "If set, reports metrics to this URL.")
 	versionFlag := flag.Bool("version", false, "Output version info")
@@ -166,7 +169,7 @@ func main() {
 	var ldb *ethdb.LDBDatabase
 
 	if *dbSupported {
-		ldb, _ = InitLDBDatabase(*ip, *port)
+		ldb, _ = InitLDBDatabase(*ip, *port, *freshDB)
 	}
 
 	host, err := p2pimpl.NewHost(&selfPeer)
